EBT: Batch messages
This is an obvious optimisation: we're buffering `IHave`s, thus may put them on
the wire as one message. Correspondingly, we can construct a batch of `Graft`s
from that. Obviously, the reply to `Graft` can now batch all payload messages
together.

A future changeset should introduce a (configurable) limit on the batch sizes.

Fixes #9